+++ src/next.c
@@ -156,6 +156,7 @@ int NextMove(struct SearchData *sd) {
         } else {
             st->st_hashmove = M_NONE;
         }
+    /* fall through */
     case GenerateCaptures: {
         BitBoard targets;
         int cnt;
@@ -203,6 +204,7 @@ int NextMove(struct SearchData *sd) {
 
         st->st_phase = GainingCapture;
     }
+    /* fall through */
     case GainingCapture:
 #ifdef VERBOSE
         Print(9, "GainingCapture\n");
@@ -231,6 +233,7 @@ int NextMove(struct SearchData *sd) {
             } else
                 break;
         }
+    /* fall through */
     case Killer1: {
         move = sd->killer->killer1;
 #ifdef VERBOSE
@@ -244,6 +247,7 @@ int NextMove(struct SearchData *sd) {
             return move;
         }
     }
+    /* fall through */
     case Killer2: {
         move = sd->killer->killer2;
 #ifdef VERBOSE
@@ -257,6 +261,7 @@ int NextMove(struct SearchData *sd) {
             return move;
         }
     }
+    /* fall through */
     case CounterMv: {
         int lmove = (p->actLog - 1)->gl_Move;
 
@@ -276,6 +281,7 @@ int NextMove(struct SearchData *sd) {
             }
         }
     }
+    /* fallthrough */
     case Killer3:
 #ifdef VERBOSE
         Print(9, "Killer3\n");
@@ -296,6 +302,7 @@ int NextMove(struct SearchData *sd) {
                 return move;
             }
         }
+        /* fallthrough */
 
     case LoosingCapture:
 #ifdef VERBOSE
@@ -324,6 +331,7 @@ int NextMove(struct SearchData *sd) {
 
             return move;
         }
+        /* fallthrough */
 
     case GenerateRest: {
 #ifdef VERBOSE
@@ -453,6 +461,7 @@ int NextEvasion(struct SearchData *sd) {
         } else {
             st->st_hashmove = M_NONE;
         }
+        /* fall through */
     case GenerateCaptures: {
         BitBoard targets;
         int cnt;
@@ -489,6 +498,7 @@ int NextEvasion(struct SearchData *sd) {
             st->st_last++;
         }
     }
+        /* fall through */
     case GainingCapture:
         while (st->st_last > st->st_first) {
             int besti = st->st_first;
@@ -516,6 +526,7 @@ int NextEvasion(struct SearchData *sd) {
             } else
                 break;
         }
+        /* fall through */
     case Killer1: {
         move = sd->killer->killer1;
 #ifdef VERBOSE
@@ -529,6 +540,7 @@ int NextEvasion(struct SearchData *sd) {
             return move;
         }
     }
+        /* fall through */
     case Killer2: {
         move = sd->killer->killer2;
 #ifdef VERBOSE
@@ -542,6 +554,7 @@ int NextEvasion(struct SearchData *sd) {
             return move;
         }
     }
+        /* fall through */
     case CounterMv: {
         int lmove = (p->actLog - 1)->gl_Move;
 
@@ -561,6 +574,7 @@ int NextEvasion(struct SearchData *sd) {
             }
         }
     }
+        /* fall through */
     case Killer3:
 #ifdef VERBOSE
         Print(9, "Killer3\n");
@@ -581,6 +595,7 @@ int NextEvasion(struct SearchData *sd) {
                 return move;
             }
         }
+        /* fall through */
     case LoosingCapture:
 #ifdef VERBOSE
         Print(9, "LoosingCapture\n");
@@ -609,6 +624,7 @@ int NextEvasion(struct SearchData *sd) {
             return move;
         }
 
+        /* fall through */
     case GenerateRest: {
         BitBoard tmp, tmp2;
         BitBoard excl;
@@ -703,6 +719,7 @@ int NextEvasion(struct SearchData *sd) {
         }
     }
 
+        /* fall through */
     case HistoryMoves:
         while (st->st_last > st->st_nc_first) {
             int besti = st->st_nc_first;
@@ -890,6 +907,7 @@ int NextMoveQ(struct SearchData *sd, int alpha) {
         GenerateQCaptures(sd, alpha);
         st->st_phase = GainingCapture;
 
+        /* fall through */
     case GainingCapture:
 #ifdef VERBOSE
         Print(9, "GainingCapture\n");